Revert a bunch of changes no longer necessary
